Finish some Testes both IT and Unit
Close #5
Close #6

+++ src/main/java/interview/challenge/restfullapinos/services/TeamServiceImpl.java
@@ -22,7 +22,7 @@ public class TeamServiceImpl implements TeamService {
 
     @Override
     public Set<Team> getTeams() {
-        log.info("Search for all Teams info");
+        log.info("Search for all Teams");
         Set<Team> teamSet = new HashSet<>();
         teamRepository.findAll().iterator().forEachRemaining(teamSet::add);
 
@@ -59,13 +59,34 @@ public class TeamServiceImpl implements TeamService {
 
     @Override
     public void deleteTeam(Team team) {
+        if (team.getId() == null)
+            throw new IllegalArgumentException("Team Id is null");
+
         teamRepository.delete(team);
         log.info("Team "+team +" was deleted!");
     }
 
     @Override
     public Team updateTeamInfo(Team team) {
-        Team savedTeam = teamRepository.save(team);
+
+        if (team.getId() == null)
+            throw new IllegalArgumentException("Team Id is null");
+
+        Team teamFromDb = teamRepository.findById(team.getId()).get();
+
+        if(team.getId().equals(teamFromDb.getId())) {
+            if (!team.getMatches().isEmpty())
+                teamFromDb.setMatches(team.getMatches());
+            if (team.getCoach()!= null)
+                teamFromDb.setCoach(team.getCoach());
+            if (team.getName()!= null)
+                teamFromDb.setName(team.getName());
+        }
+
+        if (teamFromDb.getCoach().getId()== null)
+            teamFromDb.getCoach().setId(new ObjectId().toString());
+
+        Team savedTeam = teamRepository.save(teamFromDb);
         log.info("Team "+team +" was updated!");
         return savedTeam;
     }
@@ -88,6 +109,7 @@ public class TeamServiceImpl implements TeamService {
 
         if(againstTeamObject.isPresent()) {
             Match againstMatch = new Match();
+            againstMatch.setId(new ObjectId().toString());
             againstMatch.setAgainstTeamId(teamId);
             againstMatch.setMatchResult(inverseMatchResult(match.getMatchResult()));
             Team againstTeam = againstTeamObject.get();
@@ -100,6 +122,43 @@ public class TeamServiceImpl implements TeamService {
         log.info("Match between" + teamObject.get().getName() + "and" + againstTeamObject.get().getName() + "added");
 
 
+    }
+
+    @Override
+    public void updateMatch(String teamId, Match match) {
+
+
+        if (match.getId() == null || teamId == null)
+            throw new IllegalArgumentException("Match Id or Team ID is null");
+
+        Optional<Team> teamOptional = teamRepository.findById(teamId);
+
+        if(teamOptional.isPresent()) {
+            Team team = teamOptional.get();
+            for (Match teamMatch : team.getMatches()) {
+               if (teamMatch.getId().equals(match.getId())){
+                   teamMatch.setMatchResult(match.getMatchResult());
+                   break;
+               }
+            }
+            teamRepository.save(team);
+        }
+
+        Optional<Team> againstTeamOptional = teamRepository.findById(match.getAgainstTeamId());
+
+        if (againstTeamOptional.isPresent()){
+
+            Team againstTeam = againstTeamOptional.get();
+            for (Match againstTeamMatch : againstTeam.getMatches()) {
+                if (againstTeam.getId().equals(match.getAgainstTeamId())) {
+                    againstTeamMatch.setMatchResult(inverseMatchResult(match.getMatchResult()));
+                    break;
+                }
+            }
+            teamRepository.save(againstTeam);
+        }
+
+        log.info( match +" was updated!");
     }
 
     private MatchResult inverseMatchResult(MatchResult matchResult) {
